industrial-partner-management/backend/requirements.txt
# 核心框架
fastapi==0.104.1
uvicorn[standard]==0.24.0

# 数据库与ORM
sqlalchemy==2.0.23
alembic==1.13.0
pymysql==1.1.0
aiomysql==0.2.0

# 认证与安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==41.0.7

# 数据处理
pydantic==2.5.0
pydantic-settings==2.1.0

# 工具库
python-dotenv==1.0.0
httpx==0.25.1
redis==5.0.1
celery==5.3.4

# 文件处理
Pillow==10.1.0
python-magic==0.4.27

# 开发与测试
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0

# 其他
celery==5.3.4
flower==2.0.1